requirements.txt
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.10.0
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.1
idna==2.10
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
psutil==5.9.6
py-cpuinfo==9.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.12
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
supervision==0.17.1
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.226
urllib3==2.1.0
